s3model: update path formatting logic for uploading files
After information about the type of repositories (rpm-based, deb-based)
has been added to the configuration, we can more correctly (in terms of
checks) form the path in the "_format_path" method. Before we formed
the path in accordance with the file type, now we check if this file
type is available for this repository, and then we format the path.
This approach makes the code more readable on the one hand and will
allow us to solve the problem of forming a path for uploading file with
an extension valid for several types of repositories on the other hand
(such case is not an actual problem, but it may arise in the future).