rws: add a method to update a repo metainformation
Now To update the metainformation of the repository, we can send
a post request to the address that identifies the resource.
For example:
```
curl -u test:password --request POST 127.0.0.1:5000/release/2.8/ubuntu/focal
```

Supplement:
In fact, in the case of rpm based repositories will be updated
repositories for all architextures including SRPM, for deb based
all repositories related to this distribution.