model: fix ACL handling when copy
The `Bucket.copy` method is used to load the same file into multiple
repositories. But it doesn't take into account that the ACL will be
changed to private. From documentation:
"When copying an object, you can preserve all metadata (default) or
specify new metadata. However, the ACL is not preserved and is set to
private for the user making the request."
Let's fix that.