model: update the directory check on the existing
Previously, checking for existing was done in the `_objects_to_items`
method, but if we use the `paginator` iterator when the number of
objects is a multiple of `MaxKeys`, one additional iteration will be
done and the `KeyCount` parameter in that iteration will 0 and
exception will be raised.
The existence of the directory is now checked before using
the `paginator` iterator in `get_directory` method.