dokku: change deployment mode back to `buildpack`
Dokku normally defaults to using Heroku buildpacks for deployment, but
this may be overridden by committing a valid Dockerfile to the root of
the repository and pushing the repository to the Dokku installation.

Commit 161e556 ("docker: add `Dockerfile` to build Docker images") just
added such a Dockerfile and accidentally switched the deployment to the
`Dockerfile` mode. Due to this fact, Dokku `apt` plugin on the server
started to ignore the ./apt-packages file with a needed dependency for
the app as it works only with buildpack-based applications. So we had
a wrong deployment on the server.

The solution is simple, just move ./Dockerfile to the ./docker dir.

Also, move the ./test dir to ./docker for consistency.